+++ database/migrations/2024_03_27_133432_make_bibtex_title_nullable.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        activity()->disableLogging();
+
+        Schema::table('bibliography', function (Blueprint $table) {
+            $table->text('title')->nullable()->change();
+        });
+
+        activity()->enableLogging();
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        activity()->disableLogging();
+
+        Schema::table('bibliography', function (Blueprint $table) {
+            $table->text('title')->nullable(false)->change();
+        });
+
+        activity()->enableLogging();
+    }
+};